requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
matplotlib>=3.3.0
pyperclip>=1.8.0
pillow>=8.0.0